+++ CHANGES.md
@@ -2,6 +2,16 @@
 
 Added `mask.undo()` and `mask.redo()`
 
+`setPattern()` now sets/resets the `selection`.
+
+### Breaking changes
+
+* `setPattern()` now takes an options Object for its second argument instead of
+  a String for the new value. `value` and `selection` options may be given.
+* The given (or defaulted) `selection` object is now used as-is when setting
+  initial selection. It used to be set via `setSelection()`, which moved it to
+  the first editable index.
+
 ## 1.2.0 / 2015-03-26
 
 Added customisation of format characters by passing a `formatCharacters` object